Changed the unused imports checker to ignore ignore unused imports that have a comment on the same line unless -a or --all is specified on the command line.
Tried to speed up isModule by adding a cache.  I actually did it a while ago,
probably after profiling the script on schooltool.  I do not remember if it
made any difference.  I expect it did, but the improvement was pretty small.

Originally committed 2004-02-25 18:42:17 +0200 to a different SVN repository
(python-tools) as revision 13.