Test coverage: printImports()
Added an intergation test that runs main() with the default action
(which is printImports()).

The best name for this tests was already taken by an existing test, so I
renamed it (and split it into two tests, since they were mostly
unrelated, just using the same test fixture for convenience).