Avoid ResourceWarnings and other chaff
When you open() a file but don't close it explicitly, you get a
ResourceWarning on Python 3.  It can show up in your doctest output and
make it fail.

The file.write() method returns the number of bytes written on Python 3,
and this also throws off doctest expectations.  Avoid that by assigning
to _ (used here in the "I don't care" sense).