Use pytest as a test runner
It is very unfortunate that I need to use pytest -s to run the tests.  I
was unable to find a way of avoiding that.

I haven't found a way of using a custom doctest checker with pytest
either.  The test suite might be completely broken on Windows.  This is
why I'm pushing to a branch.