Add -q/--quiet to suppress warnings
AFAIU the purpose of those 'could not find ext.mod' warnings was because
findimports was trying to understand if

    from somewhere.else import thing

was importing a class/function/variable from somewhere/else.py, or if it
was importing a thing.py module from somewhere/else/.

Most of the time I don't really care.  When I pass -N on the command
line, I _really_ don't care.